reorg: moved cookbook pages for cleanliness
